backend/pyproject.toml
# backend/pyproject.toml
[project]
name = "book-management-pwa"
version = "0.1.0"
description = "Book Management PWA Backend"
authors = [
    {name = "your-name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.12,<0.116.0",
    "uvicorn>=0.34.2,<0.35.0",
    "python-dotenv>=1.1.0,<2.0.0",
    "python-multipart>=0.0.20,<0.0.21",
    "requests>=2.31.0,<3.0.0",
    "sqlalchemy (>=2.0.41,<3.0.0)"
]

[tool.poetry.dependencies]
python = ">=3.12"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false